f32 -> f64

+++ src/main.rs
@@ -13,7 +13,7 @@ fn main() {
             .read_line(&mut weight)
             .expect("Failed to read line");
 
-        let weight: f32 = match weight.trim().parse() {
+        let weight: f64 = match weight.trim().parse() {
             Ok(num) => (num, break), //TODO: fix handling
             Err(_) => ("No a number", continue), //TODO: fix error handling 
         };
@@ -25,7 +25,7 @@ fn main() {
             .read_line(&mut increment)
             .expect("Failed to read line");
 
-        let increment: f32 = match increment.trim().parse() {
+        let increment: f64 = match increment.trim().parse() {
             Ok(num) => (num, break), //TODO: fix handling
             Err(_) => ("No a number", continue), //TODO: fix handling
         };
@@ -49,7 +49,7 @@ enum Units {
 
 //Yep, we are hardcoding a division. Specifically, a divsision such that we get a remainder, in
 //much the same way that a modulo function would.
-fn weight_division(weight: f32, increment: f32) -> (f32, f32) {
+fn weight_division(weight: f64, increment: f64) -> (f64, f64) {
     let mut result = 0.0;
     let mut remainder = weight;
     while remainder > increment {
@@ -67,7 +67,7 @@ fn weight_division(weight: f32, increment: f32) -> (f32, f32) {
 //you can get the number following the decimal (in the above example: 0.08) by doing the following:
 //(X % Y)/ Y
 //then to round down subtract this from the original division. 
-fn round_down(weight: f32, increment: f32) -> f32 {
+fn round_down(weight: f64, increment: f64) -> f64 {
     let (w, _) = weight_division(weight, increment);
     let rounded_weight = w * increment;
     rounded_weight
@@ -80,7 +80,7 @@ fn round_down(weight: f32, increment: f32) -> f32 {
 //rounding up is ceiling division
 //ex ceil(122.7/2.5) = 50
 //as above, but to always round up, you can add (1 - result) to the original division.
-fn round_up(weight: f32, increment: f32) -> f32 {
+fn round_up(weight: f64, increment: f64) -> f64 {
     let (w, r) = weight_division(weight, increment);
     if r > 0.0 {
         let rounded_weight = (w + 1.0) * increment;
@@ -95,7 +95,7 @@ fn round_up(weight: f32, increment: f32) -> f32 {
 //smart division in this case means to the nearest integer regardless of direction.
 //the rules above apply, but for smart division we would run the rounding up check, then if (1 -
 //result) is above 0.5 apply the round up rule, otherwise apply the round down down rule.
-fn smart_round(weight: f32, increment: f32) -> f32 {
+fn smart_round(weight: f64, increment: f64) -> f64 {
     let (_,r) = weight_division(weight, increment);
     let remainder_ratio = r/increment;
     if remainder_ratio > 0.5 {
